Introduce document generation (317) (22)
As pho is the first client of kiss-rdb, start to experiment with
exposing something resembling a public API there. This means

  - come up with the term “core attributes” as an improvement on
    “simple and immediate attributes”. A counterpoint to the not
    yet designed “meta” section possibly to be used for “extended”
    storage fields (like whole files), we don’t know yet. The main
    idea being “core attributes” have business meaning directly,
    immediately after they are unencoded from toml.

  - sneak in support for attribute assignments `like=this` rather
    than `like = this` because it will be a hard habit to break,
    for us and others and it is easy enough to support.

DISCUSSION: this started out as a quick-and-dirty experiment hack
because (true to the bootstrapping spirit of this project) we have
documentation in our fragments that we want to be able to read now
as a readable document. Although it’s still messy, we made good
progress that we want to take a snapshot of.

The fixture directory we add here is a snapshot of the real-life
document fragments of this project at this moment, with lots of content
removed and some content made more bland (“sanitized”).

In terms of behavior and interface, this is not yet outputting
files for lack of writing a CLI, but that will be mostly trivial.
We will want to be able to output only a single file rather than all
files for the whole collection; but keep in mind we will still need
to make the “big index” even when generating a single file, because
any arbitrary fragment can point back to another fragment, calling it
its previous or its parent.

In terms of implementation; at first we tried to “keep it simple” and
just output s-expressions (sexp’s (tuples, basically)); especially when
we started to appreciate how necessary it is to use robust markdown
parsing rather than this craziness. However, notwithstanding that
acknowledged tech debt; we are realizing that we really should make
custom classes for each kind of “symbol” in the “grammar” (each kind of
sexp).

(times)
  05-23 06:19  begin pop soft stash of big index, zz 1 hr later.
        14:25  found a difficult bug (typos the name of one dict
               for another) by overhauling the code
        16:30  first raw output of the entire document tree (just as
               lines of markdown text, no files yet.) concept is 
               proven! gotta go dance (movement!)
  05-24 17:04  decide to get a snapshot of this before big refactor.
        18:40  now