Refactor document generation (23)
Synopsis: the s-expressions got us this far (to a proof-of-concept),
but this approach has a high cost for longer-term maintainability and
future re-working. Applying the dependency inversion principle, we now
have an interface (virtually) and several classes (representing the
several kinds of AST's corresponding to the symbols of an imagined
grammar). Each class (and surrounding logical module) encapsulates
the AST-specific logic that before was sprinkled throughout the
client code.

(times)
  05-25 17:50  begin refactor
        18:10  begin experiment about how modules are loaded
        18:33  learned that from the client side, import indifferently
               and from server side, either files or __init__.py
        21:42  stop to watch live stream
  05-26 17:56  had to go party (gene farris/art park). try again
        20:53  broke it into several files, the sexp production
        20:55  take a break, maybe do laundry
        08:01  all passing
  05-27 19:02  begin for the day. about to refactor the multi-line
               block section to become one thing
        22:24  almost totally done. passing 23 tests. just need to all
               the way abstract the type out of the rendering code.
  05-28 05:23  begin again
        06:46  begin last looks