Introduce CLI for document generation. First smoke test (23)
Discussion:
The bulk of this commit is CLI code and “glue” code. We *do* introduce
one whole new magnetic (function), but all it does is a glorified but
spiritually similar version of the massive line-based test we did for
“big index” in the previous commit.

For this reason (and also the highly experimental, visual-tested nature
of things while we enter a really exploratory phase of figuring out
exactly what we want)… we do not add additional tests here.

either Hugo or the theme we are using requires that each document has
a date time (in its font matter). This is an uninteresting problem o
us so we have addressed it without coverage for now. (Ultimately it
would be great to default to a datetime derived from etc but it will
be much much later for that.)

Visual test:

    setopt no_autocd  # use this if you find that typing `pho` alone
    # does a cd into the dir with the same name as the executable doh!

    pho -v document-generate -R -o pho-doc/hugo-content/post/GENERATED -F

Small comments:

  - each fragment that was given a `document_datetime` was given one
    corresponding to its created at datetime per VCS (all one datetime).
  - oops we got too loosey-goosey with how footnoted definitions looked.
  - we added a `dry_run` debugging variable at the last minute and
    implemented the usual thing for it. to integrate it would be
    pretty trivial.

(times)

  05-28 23:17  begin CLI for doc gen
        02:57  (after one or more short naps)
               seeing our first "splay" of sub-commands
        03:41  happy with the dummy command, gonna muscle down
        04:11  our help screen is mostly ok
        04:31  begin writing the metal
        05:17  we are outputting hello world from within
               an imaginary API function
        06:12  ok we think we are really really done with
               defaulting and validating parameters
        07:26  almost done, smoke testing
        07:48  we have an actual document tree (no front matter yet)
        08:49  front matter take some work. almost done zz
  05-29 20:19  pick back up with document datetime
        22:42  visual test looks good
        23:38  now