Pre-unify "oxford join" ((899)) (38)
Synopsis: take a snapshot of "oxford join" as it is before unification

Discussion: the simple “oxford join” family of functions is implemented
redundantly in two different projects. Its disparate implementations
exercise two vastly different approaches (which is fine). Forthcoming,
we introduce “rotating buffer” which we suspect we can then use to
unify these implementations into one using a *third* approach.

(While this was not the inspiration case for “rotating buffer”, it is
a more distilled, exemplary use of our take on it for the kinds of
things we want to try with it experimentally…)

Having this snapshot will help to make clear the intention and utility
of the new way.

(1+35+31+38+73+245+141+312+23)