Parse issue identifiers (Markdown tables) (431)
Synopsis: Purely an experiment. No coverage 🙃

Discussion: Needed a programming exercise to get back in to it.
We wondered how close we are to being able to use the “dummy”
markdown storage adapter for use in our actual issues file (README.md)
to do the thing we’ve done before elsewhere (2x) and parse an issues
manifest possibly with a query to (for example) see our list of open
issues.

(times)
  05:15  begin
  07:19  refactored parsing to use an injected grammar
  13:48  now it dynamically figures out if it’s parsing issue ID’s
  12:15  now.