1/14: Introduction & Deletes
Introduction: “sakin-agac” as it existed needed to be dissolved. Its
name: a placeholder. Its purpose: enigmatic. Like the hydra of greek
myth, it had multiple heads writhing in opposing directions:

Superficially, its exit criteria was to solve the problem of how to
host a mostly static website, with some of our of homespun novelty built
in to it, and an upgrade path to “dim-sum” (currently vaporware).

But much more interesting were its detour by-products: data-producers,
the “sync” algorithm, and “tag-lyfe”. As such, we wanted to dissolve
most of “sakin-agac” and re-house most of its parts in such a way as to
be either:

  * integrated with the now supercedent “kiss-rdb” or
  * re-branded as the more obviously named “data-pipes”.

We incorrectly assumed that this would be a small-ish effort to get
started, so we thought we could get away with doing other incidental
along-the-way housekeeping, which ended up becoming the 10 (ten)
focused-topic commits that follow this one and the renames one.

Times for the whole cluster will be provided at the last commit.

These Deletes:

* `_init` (21x) and `_common_state` (1x) files obviated (see (5/14)).

* `__init__` (5x) branch nodes dissolved as S.A. branch contents get
  dissolved into other existing branch nodes.

* This “game-server” executable (1x) does nothing but call a function in a
  module. The new way (see commit (9/14)) obviates this entire file.

* `_common` (1x) was anemic and its client moved so we in-lined it.

* README from S.A. (1x) no longer relevant because the tests are
  absorbed into a different tree and numberspace.