4/14: `lazy` not `memoize` ((-2))
The old way was:
* two different functions: `memoize` and `lazy`
* we used `memoize` more than `lazy`

The new way is:
* no more two memoizing functions. get rid of the weird one. (below)
* now that the old `lazy` is gone, use the name `lazy` for old `memoize`

We removed the two tests that covered old way “lazy”.

Details:

Somehow the name `lazy` got taken up for the quite niche assembly of
defining a function in a closure and evaluating it lazily. The name
`memoize` was used for the simpler and more widespread facility of
a dereferencing function (or if you prefer, global object) that
evaluates its value lazily. (Note that the former can be made from the
latter straightforwardly, with a little less sugar.)

We decided we wanted the name `lazy` to be used for the more common
thing, because that more closely matches what a language like Swift
does (to our understanding); and we decide to do away with the other
facility entirely, to simplify our API.