7/14: `assert` not `sanity`
The old way:
* `cover_me`: a function we imported
* `sanity`

The new way:
* `cover_me` is never imported. define it in-file or just inline.
* no more `sanity` - `assert` is more idiomatic and shorter

There were a lot of files that imported the functions `sanity` and
`cover_me`. This is now seen as having value not worth the cost of
adding the dependency (and the cost of understanding our local idiom).
When we started using `sanity` all over the place, we didn’t yet know
that `assert` was built-in. Using `assert` where we use `sanity` is more
platform-idiomatic and takes less code, with the only disadvantage being
we lose the error message, but this is mitigated by comments.