Re-architect markdown table: storage (not format) adapter ((1073))
We want the old format adapter support code to go away. There were 2
format adapters that used this old code: the legacy markdown stack,
and the “producer script” stuff. In this commit we re-architect things
so that the old markdown support code is integrated into the new
storage adapter API (somewhat, but not yet all the way).

We didn’t want to get rid of the idea of “collection capabilities” but
instead refined it.

Old format adapters allowed for the format name to be passed explicitly
as a parameter. In the new way this was simplified out (opting instead
to have the convention-over-configuration of using either filename
extension or schema-file-based thing). But we now let the old parameter
bleed into the new stack because of how deeply it’s integrated in to the
old way that we are absorbing.

To shoehorn the older (but preferred) stream-based paradigm of parsing
markdown so that it co-exists next to the newer (but quick and dirty)
one, we had to introduce this idea of an “adapter variant”.

Now that they have this capability-digging method, collections need to
have some amount of reflection. We want this reflection to include both
any adapter variant and a surface name for the storage adapter.
So, “collection path” becomes “collection identity”, a bundling of
all these data points into one object.

Now that old tests deal with the new style of emissions (structured,
not “expression”); we modernize the old emission testing library to be
aware of the new shape.

(times)
  09-25 05:58  prune out non-used CLI's
  09-26 18:45  (begin previous commit)
  09-27 07:20  (end previous commit)
        07:55  resume where we were - modernize markdown format adapter
        17:34  try again after some rest
  09-29 23:56  finished a file
  09-30 04:57  now