2/2: Sunset format adapters ((1064))
DELETE NOTES

Lots of files related to sunsetting format adapters.

the “in-memory dictionary” format adapter (which existed for and lived
in tests only) gets replaced with being made to look like a producer
script.

a stream test was anemic

“normal field name via string” was anemic

lots of test support stuff wasn’t used in “tag lyfe” tests

EDIT NOTES

About losing executables [#874.5]:

We went thru a phase where we were attempting a “unix philosophy”
architecture with higher-level operations achieved by stringing together
lower-level “utilities”. In this way we ended up with lots of small
executable scripts. Although we are not abandoning this philosophy,
we are reverting to the more traditional architecture for housing the
modules. (This reversion is in part because “click” turned us on to
the `pip install`/`setup.py` way of managing executables (so our
attention drawn to how many entrypoints we have), and in part because
while unifying our collections UI up to “kiss-rdb”, we can now lose a
lot of these entry points.) NUMBER OF files lost their executability
and related code: I

Unify common kinds of context managers (pass-thru, empty, not OK).

no more “many hops”. follow the new way of importing things.

RENAME NOTES:

“legacy format adapter” becomes “entities via collection” for reasons
that are now explained in that file.

While we are doing renames, we want to go ahead and expose “kiss-rdb”’s
string scanner because it’s so much like the other one. (we almost moved
this to “script-lib” (the unofficial home of string parsing), but kiss
is by far its biggest client.)

The “kiss-rdb” string scanner *seems like*

We go ahead and promote “cheap arg parse” now to a top-level module.

“via words” not “ellipsified string via” because it’s more general and
encompasses the word wrap which is an exposed endpoint.

(times)

  09-30 05:44  begin final kill of format adaapeter
  10-05 02:33  done