Revert to virtualenv & pip from poetry ((1132))
Narrative:

We went to install the google sheets API following their instructions
(documented somewhere in this commit), and we reached a lot of
frustration at the obtusity of our using “poetry” instead of the more
de-facto standard virtualenv and pip. Because we couldn’t remember
exactly why we justified trying the experimental pip (and we didn’t
document this choice in the one file it should go it), we decided now
was a good time (indeed it felt necessary) to grok the de-facto standard
way more fully.	

The way we attempted this was brute-force, ground-up, by trying to
run our unit test suite from nothing. We re-installed (only) those
packages as revealed necessary by unit tests (🤔).

We’re not sure (and we don’t really care) but something about this
change (probably) lead to some tests failing. It may be that click
is formatting its help screens differently. (A blank line has been
added to separate multi-line items from the next item. single quotes
changed to double (or the other way), in for example “No such command”)

Anyway, this broke tests. Trying to fix the help screen tree lead to
a rewrite to try to eliminate some obtusity from a loud state-machine
approach.

Cleaning up help screen tree lead to an improvement of its parse
structure, which of course broke other tests.