Spike low-level API for timestamp syncing (45)
Also, add feature to the timestamp shell script: allow it to read one
line from STDIN as its timestamp message. This allows us to pop a line
from our “.stack” and pipe it directly in to the `ts` script (with no
other arguments) to note the completion of the item in time tracking.

(times)

  07-07 22:05  (stopped to watch anime - Berserk movie 2)
        22:20  make recording of live
        23:59  finished read from recording
  07-08 01:21  finished modifying timestamp script to read from stdin
        01:35  begin writing unit test to do sync
        03:14  begin actually writing sync code
        10:53  finally finished writing algorithm, now implement
        14:11  low level API first pass is done