Implement GUI rough pass for “update notecard” ((1184))
Visual tests has us adding parents, changing the order of children,
and (with some stumbling) editing the body, and changing the header
(and misfeature: you *can* use escape sequences in your header, more
later).

Encountered a nasty issue with newlines being escaped somewhere in our
pipeline. Worked around it for now by unescaping them.

We cleaned up our javascript enough so it feels usable by us. But it’s
the beginnings of a smell where we’d rather be using React. (That is,
of course, the whole thing with this is that we’ll use it to learn
React.)

(times)
  08-22 13:11  begin fix state stuff in UI
  08-23 08:06  progress is good but encountered bug with how it encodes
        16:39  now