Begin pho backend for delete entity ((1188))
The general thing (just now discovered in this commit) is that we want
the implementations for create, update and delete business entities
(all in front of “batch update”, today implemented in pho, hopefully one
day moved to kiss) all to be exposed to testing in the same way: by
breaking it up into two steps, the first step generating the big patch
file and the second step applying it.

To the end of testing generated edits, applying assertions against the
generated big patchfile (lines in memory) has distinct advantages over
other options:

Asserting real changes in the collection would be sluggish and code-
heavy by orders of magnitude more than it is otherwise: It could involve
making a deep, recursive copy of a filesystem tree of the fixture
collection, then the system hit of applying the patchfile to the
filesystem, then the hit of retrieving (or attempting to retrieve)
story-affected entities from the collection and making assertions over
them, and then cleaning up the filesystem tree under test.

Whereas with the big patchfiles:

- These patchfiles exist near the end of the line of a long pipeline
  of validating and interpolating/expanding an edit.

- The generated “big patchfile” is a succinct and well-suited
  representation of an atomic Unit of Work (composed of smaller units
  that all must happen “all-or-nothing”). It’s a perfect bridge between
  two useful layers of abstraction: a collection as entities in memory,
  and a collection as files on the filesystem.

- There is a rigidity and reliability to using patch files for this
  because it’s old, popular, well-tested and well-understood FOSS format

So, we went to implement delete with all this in mind. One thing that
happened was we realized that delete must modify the index file like
create does. This lead to a generalization (in some places) of the
“EIDR” (entity identifier reservation) into “IFC” (index file change),
because the same thing (but the inverse) happens on delete as on create.

Also in the spirit of unifying things for our business entity CUD,
we now expose a dry run option to all three exposures

This also led to this wild thing with parsing diffs. Dipping in to our
parser generator is terrifying and we left it a little messy.

(times)

  08-24 23:44  start delete entity (originally we wanted to do full
               vertical integration in one go: pho, GUI, backend,
               storage, all of it)
  08-25 05:58  resume
  08-26 21:07  start parse diff file
  08-27 03:06  finished parsing diff
        09:55  finished this one unit test
        13:51  now