Complete GUI core editing features ((1191))
Delete exists somewhere on the stack (we don’t remember how far down),
but for now we’re not interested in integrating it with the GUI:
  - it would be really nice to have VCS integration first, to (for
    example) ensure no data is actually lost.
  - it needs really solid “are you sure?” interface design
  - in practice we won’t delete entities often enough to warrant the
    work, in the near term

Getting a “create” was of course more involved than we first expected:
Because in practice it will be this way (so we thought, before we
thought about it), we figured you would typically be creating a new
notecard “off” of an existing parent notecard. (In fact maybe it is
more often you will do it “next” off of a previous notecard.)

So instead of plain old “create” there is “add child”. All sorts of
small details cropped up: When creating new, you don’t want to see the
“identifier” field. (You don’t have an identifier until you succeed in
creating.) Seeing the blank identifier field there feels weird. The form
flows better with extraneous fields not shown.

Because you’re a child off of a parent, you also can’t have a previous.
(You can only have one between parent and previous.). So ditto we don’t
want that field there.

Creating a child (the button) is only exposed when you’re editing node.
So when you cancel out of the child form it makes sense to “pop back”
to the parent you were in mid-edit of. (This “pop back” behavior doesn’t
work as expected in a couple ways: Form values might not be sticky, and
after successful creation of child, this (imaginary) “stack” is
collapsed.These are both seen as way out of scope issues for now.)

So what happened was we turned the form into a more rigid state machine
with four states:

  BEGINNING -> VIEWING <-> EDITING <-> CREATING
                  ^———————————————————————/

You can’t re-enter BEGINNING after you leave it.

You can go to edit an existing entity (update) and cancel, and this pops
you back from editing to viewing.

You can go to add a child and then cancel, and this pops you back in to
the editing state from creating.

After successful create, you pop back down into viewing, but with that
selfsame new entity you just created.

Form values from any forms in mid-edit that end up below you on a stack
are lost when you pop another frame over that form.

Other points:
  - unify some names. states all get “-ing” for now.
  - Sure it would be nice to have GUI coverage, but we have to
    bootstrap that in somewhere.

(times)
  08-28 02:03  begin GUI for create
        02:53  progress good but zz
        23:08  done except for tiny issues with disappearing things
  08-29 15:07  visited family earlier. now.