( 4/16) Conventionalize whitespace & formatting
For whatever reason, when we started in Python, we were obsessed with
representing list-like literals (dictionaries, tuples, lists, import
lists) in such a way that item-level changes could be seen as line-level
changes. This lead to lists that were preternaturally vertical, and lots
of lines consisting of only a close parenthesis.

The sands of time have shown us that the return-on-investment in this
convention is pretty low; that is, we don't change the composition of
our lists frequently enough to justify this strnage-looking style.